reorganizing the mask handling, update of interface files
