+++ ecmean/interfaces/interface_GLOBO.yml
@@ -28,18 +28,23 @@ filetype:
     dir: ''
     component: globo
 
-# Specify additional properties of each component
-# such as inifiles, an extra help file containing 
-# info such as areas, used in set_gridfixes
-# inifile can be formatted with: expname
+# Specify additional inifiles  of each component
+# maskfile: where the land-sea mask is, fundamental for atmospheric component 
+# gridfile: where the properties of the grid can be extracted, can be a standard file
+# areafile: where the grid cell area is
+# please not that for both grid area computation and itnerpolation weight computation
+# the code will look in priority from areafile to gridfile to maskfile. 
+# inifile can be formatted with all the possible keywords listed above
 # If a relative path is used this is attached to $ECEDIR/$basedir
 component:
   globo:
-    inifile: 'GLOBONE_BCs.nc'
-    atmfile: 'GLOBONE_atm_6hrs_{year1}.nc'
+    maskfile: 'GLOBONE_BCs.nc'
+    gridfile: 'GLOBONE_atm_6hrs_{year1}.nc'
+    areafile: ''
   fesom:
-    inifile: None
-    areafile: None
+    maskfile: ''
+    gridfile: ''
+    areafile: ''
 
 # Definition of individual variables
 variables:
