--- tests/table/PI4_RK08_historical_EC-Earth3_r1i1p1f1_1990_1990.yml
@@ -1,24 +1,0 @@
-pr:
-  ALL:
-    Global: 38.218
-    North Midlat: 23.301
-    Tropical: 30.225
-    South Midlat: 67.543
-ta:
-  ALL:
-    Global: 22.657
-    North Midlat: 17.383
-    Tropical: 8.335
-    South Midlat: 56.836
-tos:
-  ALL:
-    Global: 37.626
-    North Midlat: 22.567
-    Tropical: 14.62
-    South Midlat: 92.944
-siconc:
-  ALL:
-    Global: 37.767
-    North Midlat: 60.389
-    Tropical: .nan
-    South Midlat: 24.388
